+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- fix(wizard): Handle missing auth token in wizard API endpoint response (#566)
+
 ## 3.22.0
 
 - feat(nextjs): Ask users about tunnelRoute option (#556)